+++ packages/service-content/src/lib/professors/queries/get-professor-courses.ts
@@ -54,7 +54,7 @@ export const getProfessorCoursesQuery = ({
       c.remaining_seats,
       c.number_of_rating,
       c.sum_of_all_rating,
-      COALESCE(NULLIF(c.sum_of_all_rating, 0) / NULLIF(c.number_of_rating, 0), 0) AS average_rating,
+      COALESCE(NULLIF(c.sum_of_all_rating::float, 0) / NULLIF(c.number_of_rating, 0), 0) AS average_rating,
       cl.name,
       cl.goal,
       cl.objectives,
